Move locales.ts to new folder /localization (placed directly under /client)
