+++ src/client/localization/locales.ts
